entrada de caracteres

+++ aulas/aula 3/entrada-de-dados.c
@@ -19,18 +19,18 @@
 
 int main()
 {
-    float numero1, numero2, soma;
+    int numero1, numero2, soma;
 
     printf("informe o primeiro numero:");
-    scanf("%f", &numero1);
+    scanf("%d", &numero1);
 
     printf("informe o segundo numero:");
-    scanf("%f", &numero2);
+    scanf("%d", &numero2);
 
     soma = numero1 + numero2;
 
     //10 + 20 = 30
-    printf("%.2f + %.2f = %.2f\n", numero1, numero2, soma);
+    printf("%d + %d = %d\n", numero1, numero2, soma);
 
     system("pause");
     return 0;
